+++ bitframe-authentication/core/src/commonMain/kotlin/bitframe/authentication/signin/Session.kt
@@ -7,24 +7,26 @@ import bitframe.authentication.spaces.Space
 import bitframe.authentication.users.User
 import kotlin.js.JsExport
 
-sealed class Session {
-    object Unknown : Session()
+sealed class Session<out T> {
+    object Unknown : Session<Nothing>()
 
-    class SignedIn(
+    data class SignedIn<out T>(
         val app: App,
         val space: Space,
         val user: User,
-    ) : Session()
+        val data: T
+    ) : Session<T>()
 
-    class Conundrum(
+    data class Conundrum(
         val app: App,
         val spaces: List<Space>,
         val user: User
-    ) : Session()
+    ) : Session<Nothing>()
 
-    class SignedOut(
+    data class SignedOut<out T>(
         val app: App,
         val space: Space?,
         val user: User?,
-    ) : Session()
+        val data: T?
+    ) : Session<T>()
 }