+++ bitframe-authentication/services/core/src/commonMain/kotlin/bitframe/authentication/signin/SignInServiceImpl.kt
@@ -4,6 +4,7 @@ import bitframe.authentication.AuthenticationDaoProvider
 import bitframe.authentication.apps.App
 import bitframe.authentication.spaces.Space
 import bitframe.authentication.spaces.SpacesDao
+import bitframe.authentication.users.User
 import bitframe.authentication.users.UsersDao
 import bitframe.daos.conditions.contains
 import bitframe.service.config.ServiceConfig
@@ -11,16 +12,14 @@ import later.Later
 import later.await
 import later.later
 
-class SignInServiceImpl(
+open class SignInServiceImpl(
     private val usersDao: UsersDao,
     private val spacesDao: SpacesDao,
-    private val config: ServiceConfig
-) : SignInService() {
+    override val config: ServiceConfig
+) : SignInService<Nothing?>() {
     constructor(provider: AuthenticationDaoProvider, config: ServiceConfig) : this(provider.users, provider.spaces, config)
 
-    private val scope = config.scope
-    override fun signIn(credentials: SignInCredentials): Later<LoginConundrum> = scope.later {
-        validate(credentials)
+    override fun executeSignIn(credentials: SignInCredentials): Later<LoginConundrum> = scope.later {
         val matches = usersDao.all(where = "contacts" contains credentials.alias).await()
         if (matches.isEmpty()) throw RuntimeException("User with loginId=${credentials.alias}, not found")
         val match = matches.first()
@@ -31,24 +30,12 @@ class SignInServiceImpl(
             session.value = if (it.spaces.size > 1) {
                 Session.Conundrum(App(config.appId), it.spaces, it.user)
             } else {
-                Session.SignedIn(App(config.appId), it.spaces.first(), it.user)
+                makeSession(App(config.appId), it.spaces.first(), it.user).await()
             }
         }
     }
 
-    override fun resolve(space: Space): Later<Session.SignedIn> = scope.later {
-        val error = IllegalStateException(
-            """
-                You are attempting to resolve a non exiting conundrum,
-                
-                Make sure you have tried to signIn and the result obtained was a LoginConundrum with more that one space
-                """.trimIndent()
-        )
-        when (val s = session.value) {
-            Session.Unknown -> throw error
-            is Session.SignedIn -> Session.SignedIn(App(config.appId), space, s.user)
-            is Session.Conundrum -> Session.SignedIn(App(config.appId), space, s.user)
-            is Session.SignedOut -> throw error
-        }.also { session.value = it }
+    override fun makeSession(app: App, space: Space, user: User) = scope.later {
+        Session.SignedIn(app, space, user, null)
     }
 }