+++ pi-monitor/pi-monitor-client/browser/react/src/jvmTest/kotlin/acceptance/authentication/SignUp.kt
@@ -3,15 +3,11 @@
 package acceptance.authentication
 
 import acceptance.utils.AcceptanceTest
-import expect.expect
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.TestInstance
 import org.testcontainers.junit.jupiter.Testcontainers
-import pimonitor.authentication.signup.IndividualRegistrationParams
-import pimonitor.authentication.signup.MonitorBusinessParams
-import pimonitor.screens.api.toBeVisible
+import pimonitor.monitors.SignUpParams
 import pimonitor.test
-import kotlin.test.Ignore
 import kotlin.test.Test
 
 @Testcontainers
@@ -21,34 +17,31 @@ class SignUp : AcceptanceTest() {
     @Nested
     inner class `Individual Registration` {
         // Given a Person with details
-        private val person = IndividualRegistrationParams(
+        private val person = SignUpParams.Individual(
             name = "John Doe", email = "john.doe@johndoeinc.com", password = "1234"
         )
 
         @Test
         fun should_should_be_able_to_sign_individually() = application.test {
             val signUpScreen = openSignUpScreen()
-            signUpScreen.signUpIndividuallyAs(person)
+            signUpScreen.signUp(person)
             signUpScreen.expectUserToBeRegistered()
         }
     }
 
     @Nested
     inner class `Organisational Registration` {
-        // Given
-        private val person = IndividualRegistrationParams(
-            name = "John Doe", email = "john.doe@johndoeinc.com", password = "1234"
-        )
-
-        // Given
-        private val business = MonitorBusinessParams(
-            name = "John Doe Inc.", email = "support@johndoeinc.com"
+        private val params = SignUpParams.Business(
+            businessName = "John Doe Inc.",
+            individualName = "John Doe",
+            individualEmail = "john.doe@johndoeinc.com",
+            password = "1234"
         )
 
         @Test
         fun should_register_a_new_user() = application.test {
             val signUpScreen = openSignUpScreen()
-            signUpScreen.signUpAs(person, representing = business)
+            signUpScreen.signUp(with = params)
             signUpScreen.expectUserToBeRegistered()
         }
     }