+++ pi-monitor/pi-monitor-client/sdks/core/src/jsMain/kotlin/pimonitor/authentication/signup/exports/ParamsUtils.kt
@@ -1,11 +1,6 @@
 package pimonitor.authentication.signup.exports
 
-import pimonitor.authentication.signup.IndividualRegistrationParams
-import pimonitor.authentication.signup.MonitorBusinessParams
 import pimonitor.monitors.SignUpParams
-
-fun RegisterIndividualParams.toIndividualRegistrationParams() = IndividualRegistrationParams(name, email, password)
-fun RegisterOrganisationParams.toMonitorBusinessParams() = MonitorBusinessParams(name, email)
 
 fun RegisterIndividualParams.toSignUpParams() = SignUpParams.Individual(name, email, password)