+++ pi-monitor/pi-monitor-client/sdks/full/src/jsMain/kotlin/pimonitor/evaluation/business/BusinessesScope.kt
@@ -3,12 +3,12 @@
 
 package pimonitor.evaluation.business
 
-import pimonitor.evaluation.businesses.BusinessService
+import pimonitor.evaluation.businesses.BusinessesService
 import viewmodel.ViewModel
 import pimonitor.evaluation.business.BusinessesIntent as Intent
 import pimonitor.evaluation.business.BusinessesState as State
 
-class BusinessesScope(service: BusinessService) {
+class BusinessesScope(service: BusinessesService) {
     val viewModel: ViewModel<Intent, State> = BusinessViewModel(service)
 
     val loadBusiness: () -> Unit = { viewModel.post(Intent.LoadBusinesses) }