+++ pi-monitor/pi-monitor-server/src/main/kotlin/pimonitor/PiMonitorServer.kt
@@ -3,11 +3,14 @@ package pimonitor
 import bitframe.Application
 import bitframe.daos.config.InMemoryDaoConfig
 import bitframe.server.modules.Module
-import bitframe.server.modules.authentication.AuthenticationService
 import bitframe.server.modules.authentication.AuthenticationModuleImpl
+import bitframe.server.modules.authentication.AuthenticationService
 import bitframe.service.config.ServiceConfig
 import pimonitor.authentication.signup.SignUpController
 import pimonitor.authentication.signup.SignUpModule
+import pimonitor.monitored.MonitoredBusiness
+import pimonitor.monitors.CooperateMonitor
+import pimonitor.monitors.IndividualMonitor
 import pimonitor.monitors.MonitorDaoInMemory
 import java.io.File
 
@@ -25,8 +28,8 @@ fun PiMonitorServer(
                 service = authService.users
             )
         ),
-        Module<Monitor>(),
-        Module<Monitor.Business>("monitor-businesses"),
-        Module<Monitored>(),
+        Module<IndividualMonitor>(),
+        Module<CooperateMonitor>("monitor-businesses"),
+        Module<MonitoredBusiness>(),
     )
 )