+++ pi-monitor/pi-monitor-services/core/src/commonMain/kotlin/pimonitor/PiMonitorService.kt
@@ -6,12 +6,14 @@ import bitframe.BitframeService
 import bitframe.authentication.signin.SignInService
 import bitframe.authentication.users.UsersService
 import pimonitor.authentication.signup.SignUpService
-import pimonitor.evaluation.businesses.BusinessService
+import pimonitor.evaluation.businesses.BusinessesService
+import pimonitor.monitors.MonitorsService
 import kotlin.js.JsExport
 
 abstract class PiMonitorService(
     override val users: UsersService,
-    override val signIn: SignInService,
+    override val signIn: SignInService<*>,
     val signUp: SignUpService,
-    val businesses: BusinessService,
+    val monitors: MonitorsService,
+    val businesses: BusinessesService,
 ) : BitframeService