+++ pi-monitor/pi-monitor-services/core/src/commonMain/kotlin/pimonitor/evaluation/businesses/BusinessServiceImpl.kt
@@ -5,7 +5,7 @@ import pimonitor.monitored.MonitoredBusiness
 import pimonitor.monitored.MonitoredBusinessDao
 
 class BusinessServiceImpl(
-    private val dao: MonitoredBusinessDao
-) : BusinessService() {
+    private val dao: MonitoredBusinessDao,
+) : BusinessesService() {
     override fun all(): Later<List<MonitoredBusiness>> = dao.all()
 }